backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2
pydantic>=2
pydantic-settings>=2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
# Choose one of these depending on DB:
psycopg2-binary      # PostgreSQL
# mysqlclient        # MySQL/MariaDB
# If you just want to try quickly without installing a server, use SQLite (no extra driver)
